requirements.txt
pandas==2.2.2
requests==2.31.0
openpyxl==3.1.2
mysql-connector-python==8.3.0
scikit-learn==1.4.2
SQLAlchemy==2.0.30
markdown==3.5.2
weasyprint==60.2
